requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.3
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.2.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
